adicionado validação na exclusão de produto

+++ src/controllers/conta.js
@@ -24,7 +24,12 @@ const cadastrarProduto = async (req, res) => {
   const { descricao, quantidade_estoque, valor, categoria_id } = req.body;
 
   try {
-    const categoriaExistente = await dataExistente('categorias', 'id', '=', categoria_id);
+    const categoriaExistente = await dataExistente(
+      "categorias",
+      "id",
+      "=",
+      categoria_id
+    );
     if (!categoriaExistente) {
       return res.status(400).json(chat.error400);
     }
@@ -34,7 +39,7 @@ const cadastrarProduto = async (req, res) => {
         descricao,
         quantidade_estoque,
         valor,
-        categoria_id
+        categoria_id,
       })
       .returning("*");
 
@@ -50,17 +55,22 @@ const editarProduto = async (req, res) => {
   const { descricao, quantidade_estoque, valor, categoria_id } = req.body;
 
   if (!id) {
-    return res.status(404).json(chat.error404)
+    return res.status(404).json(chat.error404);
   }
 
   try {
-    const produtoExistente = await dataExistente('produtos', "id", '=', id)
+    const produtoExistente = await dataExistente("produtos", "id", "=", id);
 
     if (categoria_id) {
-      const categoriaExistente = await dataExistente('categorias', "id", "=", categoria_id);
+      const categoriaExistente = await dataExistente(
+        "categorias",
+        "id",
+        "=",
+        categoria_id
+      );
 
       if (!categoriaExistente) {
-        return res.status(400).json(chat.error400)
+        return res.status(400).json(chat.error400);
       }
     }
 
@@ -74,7 +84,7 @@ const editarProduto = async (req, res) => {
       descricao: descricao || produto.descricao,
       quantidade_estoque: quantidade_estoque || produto.quantidade_estoque,
       valor: valor || produto.valor,
-      categoria_id: categoria_id || produto.categoria_id
+      categoria_id: categoria_id || produto.categoria_id,
     };
 
     if (
@@ -83,7 +93,7 @@ const editarProduto = async (req, res) => {
       produtoAtualizado.valor === produto.valor &&
       produtoAtualizado.categoria_id === produto.categoria_id
     ) {
-      return res.status(304).json()
+      return res.status(304).json();
     }
 
     await knex("produtos").where({ id }).update(produtoAtualizado);
@@ -96,12 +106,11 @@ const editarProduto = async (req, res) => {
 
 const listarProdutos = async (req, res) => {
   try {
-    const produtos = await knex('produtos');
+    const produtos = await knex("produtos");
     if (produtos.length < 1) {
       return res.status(404).json(chat.error404);
     }
-    return res.status(200).json(produtos)
-
+    return res.status(200).json(produtos);
   } catch (error) {
     return res.status(500).json(chat.error500);
   }
@@ -110,13 +119,12 @@ const listarProdutos = async (req, res) => {
 const DetalharProduto = async (req, res) => {
   const { id } = req.params;
   try {
-    const produtoExistente = await dataExistente('produtos', 'id', '=', id);
+    const produtoExistente = await dataExistente("produtos", "id", "=", id);
 
     if (produtoExistente.length < 1) {
       return res.status(404).json(chat.error404);
     }
-    return res.status(200).json(produtoExistente[0])
-
+    return res.status(200).json(produtoExistente[0]);
   } catch (error) {
     return res.status(500).json(chat.error500);
   }
@@ -126,70 +134,77 @@ const deletarProduto = async (req, res) => {
   const { id } = req.params;
 
   if (!id) {
-    return res.status(404).json(chat.error404)
+    return res.status(404).json(chat.error404);
   }
 
   try {
-    const produtoExistente = await dataExistente('produtos', 'id', '=', id);
+    const produtoExistente = await dataExistente("produtos", "id", "=", id);
 
     if (produtoExistente.length < 1) {
-      return res.status(404).json(chat.error404)
+      return res.status(404).json(chat.error404);
     }
 
-    await knex('produtos').delete().where({ id });
+    const produtoVinculadoEmPedido = await knex("pedido_produtos")
+      .where("produto_id", "=", id)
+      .select("pedido_id");
+
+    if (produtoVinculadoEmPedido.length > 0) {
+      return res.status(404).json(chat.error403);
+    }
+
+    await knex("produtos").delete().where({ id });
 
     return res.status(204).json();
-
   } catch (error) {
     console.log(error.message);
-    return res.status(500).json(chat.error500)
+    return res.status(500).json(chat.error500);
   }
 };
 
 const cadastrarCliente = async (req, res) => {
-  const { nome, email, cpf, cep, rua, numero, bairro, cidade, estado } = req.body;
+  const { nome, email, cpf, cep, rua, numero, bairro, cidade, estado } =
+    req.body;
 
   try {
+    const emailExistente = await dataExistente("clientes", "email", "=", email);
 
-    const emailExistente = await dataExistente('clientes', 'email', '=', email);
-
-    const cpfExistente = await dataExistente('clientes', 'cpf', '=', cpf);
+    const cpfExistente = await dataExistente("clientes", "cpf", "=", cpf);
 
     if (cpfExistente.length > 0 || emailExistente.length > 0) {
       return res.status(400).json(chat.error400);
     }
 
-    const novoCliente = await knex('clientes').insert({
-      nome,
-      email,
-      cpf,
-      cep,
-      rua,
-      numero,
-      bairro,
-      cidade,
-      estado
-    })
-      .returning('*');
+    const novoCliente = await knex("clientes")
+      .insert({
+        nome,
+        email,
+        cpf,
+        cep,
+        rua,
+        numero,
+        bairro,
+        cidade,
+        estado
+      })
+      .returning("*");
 
-    return res.status(201).json(novoCliente[0])
-
+    return res.status(201).json(novoCliente[0]);
   } catch (error) {
     return res.status(500).json(chat.error500);
   }
-
-}
+};
 
 const editarDadosDoCliente = async (req, res) => {
   const { id } = req.params;
-  const { nome, email, cpf, cep, rua, numero, bairro, cidade, estado } = req.body;
+  const { nome, email, cpf, cep, rua, numero, bairro, cidade, estado } =
+    req.body;
 
   if (Object.keys(req.body).length === 0) {
-    return res.status(400).json(chat.error400)
+    return res.status(400).json(chat.error400);
   }
 
   try {
-    const clienteExistente = await dataExistente('clientes', 'id', '=', id);
+    const clienteExistente = await dataExistente("clientes", "id", "=", id);
 
     if (!clienteExistente) {
       return res.status(400).json(chat.error400);
@@ -198,9 +213,9 @@ const editarDadosDoCliente = async (req, res) => {
     const cliente = clienteExistente[0];
 
     if (email) {
-      const emailExistente = await knex('clientes')
-        .where('email', '=', email)
-        .where('id', '<>', id)
+      const emailExistente = await knex("clientes")
+        .where("email", "=", email)
+        .where("id", "<>", id)
         .first();
 
       if (emailExistente) {
@@ -209,9 +224,9 @@ const editarDadosDoCliente = async (req, res) => {
     }
 
     if (cpf) {
-      const cpfExistente = await knex('clientes')
-        .where('cpf', '=', cpf)
-        .where('id', '<>', id)
+      const cpfExistente = await knex("clientes")
+        .where("cpf", "=", cpf)
+        .where("id", "<>", id)
         .first();
 
       if (cpfExistente) {
@@ -245,10 +260,9 @@ const editarDadosDoCliente = async (req, res) => {
       return res.status(304).json();
     }
 
-    await knex('clientes').where({ id }).update(clienteAtualizado);
+    await knex("clientes").where({ id }).update(clienteAtualizado);
 
     return res.status(200).json(chat.status200);
-
   } catch (error) {
     console.log(error.message);
     return res.status(500).json(chat.error500);
@@ -256,31 +270,28 @@ const editarDadosDoCliente = async (req, res) => {
 };
 
 const listarClientes = async (req, res) => {
-
   try {
-    const clientes = await knex('clientes');
+    const clientes = await knex("clientes");
     if (clientes.length < 1) {
       return res.status(404).json(chat.error404);
     }
-    return res.status(200).json(clientes)
-
+    return res.status(200).json(clientes);
   } catch (error) {
     return res.status(500).json(chat.error500);
   }
-}
+};
 
 const detalharCliente = async (req, res) => {
   const { id } = req.params;
 
   try {
-    const clienteExistente = await dataExistente('clientes', 'id', '=', id);
+    const clienteExistente = await dataExistente("clientes", "id", "=", id);
 
     if (clienteExistente.length < 1) {
-      const clientes = await knex('clientes')
-      return res.status(200).json(clientes)
+      const clientes = await knex("clientes");
+      return res.status(200).json(clientes);
     }
-    return res.status(200).json(clienteExistente[0])
-
+    return res.status(200).json(clienteExistente[0]);
   } catch (error) {
     return res.status(500).json(chat.error500);
   }
@@ -298,7 +309,12 @@ const cadastrarPedido = async (req, res) => {
       return res.status(400).json(chat.error400);
     }
 
-    const clienteExistente = await dataExistente('clientes', 'id', '=', cliente_id);
+    const clienteExistente = await dataExistente(
+      "clientes",
+      "id",
+      "=",
+      cliente_id
+    );
 
     const cliente = clienteExistente[0];
 
@@ -311,21 +327,26 @@ const cadastrarPedido = async (req, res) => {
     for (const produtoInfo of pedidos_produtos) {
       const { produto_id, quantidade_produto } = produtoInfo;
 
-      let produto = await knex('produtos').where({ id: produto_id }).first();
+      let produto = await knex("produtos").where({ id: produto_id }).first();
 
       if (!produto) {
-        return res.status(404).json({ error: 'Produto não encontrado' });
+        return res.status(404).json({ error: "Produto não encontrado" });
       }
 
       if (produto.quantidade_estoque < quantidade_produto) {
-        return res.status(400).json({ error: `Quantidade em estoque insuficiente para o produto ${produto.descricao}` });
+        return res
+          .status(400)
+          .json({
+            error: `Quantidade em estoque insuficiente para o produto ${produto.descricao}`,
+          });
       }
 
       valorTotal += quantidade_produto * produto.valor;
 
-      await knex('produtos').where({ id: produto_id }).decrement('quantidade_estoque', quantidade_produto);
+      await knex("produtos")
+        .where({ id: produto_id })
+        .decrement("quantidade_estoque", quantidade_produto);
     }
-
     await knex('pedidos').insert({
       cliente_id,
       observacao,
@@ -347,10 +368,9 @@ const cadastrarPedido = async (req, res) => {
     )
 
     return res.status(201).json();
-
   } catch (error) {
     console.log(error.message);
-    return res.status(500).json(chat.error500)
+    return res.status(500).json(chat.error500);
   }
 };
